Upgrade to Scala.js v0.6.33.
That is the last version that supports Scala 2.10.2.

+++ html5-graphics/src/main/scala/com/funlabyrinthe/graphics/html/CanvasWrapper.scala
@@ -5,11 +5,11 @@ import com.funlabyrinthe.core.graphics._
 import org.scalajs.dom
 
 class CanvasWrapper(val delegate: dom.HTMLCanvasElement) extends Canvas {
-  def width: Double = delegate.width
-  def width_=(value: Double): Unit = delegate.width = value
+  def width: Double = delegate.width.toDouble
+  def width_=(value: Double): Unit = delegate.width = value.toInt
 
-  def height: Double = delegate.height
-  def height_=(value: Double): Unit = delegate.height = value
+  def height: Double = delegate.height.toDouble
+  def height_=(value: Double): Unit = delegate.height = value.toInt
 
   private lazy val graphicsContext2D = {
     new GraphicsContextWrapper(
